+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # develop
+  * Feature: Add `#say_characters`, which uses `SayDigits` to override adhearsion's built in SSML-generating method
 
 # v1.4.0
   * Feature: Add `#goto`, which will pass a call to another context/extension/priority and disconnect from Adhearsion